requirements.txt
prefect
requests
beautifulsoup4
pytest
pytest-cov
faiss-cpu
